+++ backends/asdf/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+#### 🐞 Fixes
+
+- Fixed some issues when the plugin is used as "asdf" directly.
+
 ## 0.1.0
 
 #### 🎉 Release